+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Added support for diagrams.
 * statemachine integrity checks are performed at import time instead of when
   instance creation. This allows early feedback on invalid definitions.
+* Drop support for adding `validators` as assigment to `Event`.
 
 
 0.9.0 (2022-12-21)